Forcing backwards compatibility using Legacy and PHP8 classes
